Move RobotsAllowAllShortUrlsConfigOption to dedicated namespace
