+++ composer.json
@@ -17,10 +17,10 @@
         "laminas/laminas-config-aggregator": "^1.15",
         "laminas/laminas-servicemanager": "^3.22",
         "laminas/laminas-stdlib": "^3.19",
-        "shlinkio/shlink-config": "^3.0 || ^2.5",
-        "symfony/console": "^7.0 || ^6.4",
-        "symfony/filesystem": "^7.0 || ^6.4",
-        "symfony/process": "^7.0 || ^6.4"
+        "shlinkio/shlink-config": "^3.0",
+        "symfony/console": "^7.1",
+        "symfony/filesystem": "^7.1",
+        "symfony/process": "^7.1"
     },
     "require-dev": {
         "devster/ubench": "^2.1",
@@ -29,7 +29,7 @@
         "phpunit/phpunit": "^11.1",
         "roave/security-advisories": "dev-master",
         "shlinkio/php-coding-standard": "~2.3.0",
-        "symfony/var-dumper": "^7.0 || ^6.4"
+        "symfony/var-dumper": "^7.1"
     },
     "autoload": {
         "psr-4": {
